[IOT-789] [IOT-792] [IOT-794] [IOT-1763 partial fix] Remove AMS and clean up SRM context and response code.
The AMS code is dead, broken, and needs to be removed.

Also, the SRM uses a tangled and inconsistent set of context objects,
leaving it unclear which request is being processed, what the state
of processing is, and whether or not a response has been sent to the
requester.

This patch fixes this and should be used as the basis
for future improvements to the policy engine (ACL) code,
SVR handling, etc.

There is some additional cleanup I'd like to do, but this needs
to get merged, and is already a large change, so I'll add those
other fixes as smaller patches once this is merged to master.

NOTE TO REVIEWERS: a few places are marked "TODO Samsung" and need
review by Samsung engr because it is a Samsung feature.

patch set 7: deleted commented out code and files

patch set 8: fixed const assignment warning that was causing Arduino
build to fail

patch set 9: rebased and merged 16883

patch set 10: addressed Kevin's comments and automated build issue

patch set 11: automated build issue

patch set 12: automated build issue

Change-Id: I9560c8aeb1ee59925c0b7aa1282f0cf39f3cae9b
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16719
Reviewed-by: Kevin Kane <kkane@microsoft.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>