Fix unused-const-variable compiler warning
The variable MSECS_PER_SEC was declared but not used.
There were two ways to solve this compiler warning.
  1. Delete variable
  2. Use variable

Option 2 was used for this change. In several locations that the
hardcoded number 1000 was used that was clearly intended to be
MSECS_PER_SEC. This also results in a few locations where
multiplication by MSECS_PER_SEC is used to conver the time to
micro seconds.

While doing this one location was updated to also use the USECS_PER_SEC
const variable.

Change-Id: I448fc1ca614b0d8be7ea4836da6406dbd69ef325
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16801
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>