IOT-1789 Eliminate duplicate messages regardless of addr family
There can be duplicate messages coming from multiple ipv6 (or v4)
addresses in the system. The presence of ipv6 link-local for instance
could cause duplicate messages to appear. They can also occur due to
retransmits.

This change extends the existing mechanism of looking at the previous
four unique messages and dropping the new one if the COAP token
matched any of the four messages regardless of the address family.

Change-Id: I88617ad441f51a730cdf75f8a0b01c01ee002eaa
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16973
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: jihwan seo <jihwan.seo@samsung.com>
Reviewed-by: Mike Fenelon <mike.fenelon@microsoft.com>