Fix for missing-field-initializers compiler warning
The OCDevAddr struct has 7 values only 2 were
initilized when the addr struct was created.

Two values the port and addr were set after the
initilization. All values are now intilized when
the OCDevAddr struct is created.

In addition the EXPECT_EQ was changed to EXPECT_STREQ
the macro undertands that a string is expected and
will display more informative errors if the test fails.

Change-Id: I74093d0a42961f65c3709254e949b8feb8aec587
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17101
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>