Add to checking logic to set subject of cred in case of NULL or *
Issue : If deviceuuid is null and required pre-install certificate in SVR DB,
        it is impossible to set a subject of SIGNED_ASYMMETRIC_KEY type (certificate).

According to OCF Security Spec and Raml, subject of Cred allows UUID format only,
so CTT decides a failure in case of wildcard(*).
But, it is not clear how to set a subject of cred in case of SIGNED_ASYMMETRIC_KEY on spec.

So, this patch set it with own deviceuuid in case of NULL or wildcard,
because own deviceuuid is only UUID a device knows at all time.

Patch 1,2: Upload patch
Patch 3 : Set subject of cred with own deviceuuid in case of NULL or wildcard
Patch 4: Apply review comment
Patch 5-8 : Re-trigger jenkins build

Change-Id: Idb09357a821be81bb9f05489ebb40403e66c3514
Signed-off-by: js126.lee <js126.lee@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17153
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Kevin Kane <kkane@microsoft.com>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>