Fixed documentation warnings for c-docs
Fixed all of the documentation warnings found when running
Doxygen for the C documentation.

All of the documentation warnings fell into 3 areas:
1. Error using the @param[in,out] syntax
2. @param did not match the actual parameter name
3. Paramerter was missing documentation all together

For  case 3 an attempt to provide the best documentation
was done based on context. In a few situations the parameter
documentation is just restating the type of parameter being
passed in.

Change-Id: I06797a471dcefaaa2d9b50dc44ca4228cfafb941
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17543
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>