Fixed documentation warnings for cpp-docs
Fixed all of the documentation warnings found when running
Doxygen for C++ documentation.

Most of the warnings were due to @param tag not matching
the actual param.

For the copydoc warning foun in OCPlatform the @copydoc
tag was removed because it was not properly copying the
documentation. No clear fix other than actually copying
the documentation.

Change-Id: Ic861ff7dbeb9c7f17d7f39dace05da92f3bb43d1
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17569
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>