csdk: cleanup compiler warnings in occlient and occlientcoll
warnings -Wconversion-null and -Wmissing-field-initializers were
caused by trying to set a char array to NULL. Its now set to an
empty string. It is not removed so the entry count in the queryInterface
is the same as the number of test casses. The value is no longer used out
of the queryInterface.

If the TEST_GET_NULL test case is used it is special cased to use the
new InitGetRequestOnNullResource.

OC_UNUSED macro has been added to prevent the -Wunused-parameter warning.

Change-Id: I4a32e663bca574a6735f3efad2de42046328cc57
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17749
Reviewed-by: Joseph Morrow <joseph.l.morrow@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>