connectivity: Fix discovery+blockwise regression
A regression has been reported on discovery when large payloads are used.

After investigation in blockwise feature,
it appeared that regression occurred with this change:

https://gerrit.iotivity.org/gerrit/#/c/17419/

Author assumed that payload should fit into CoAP PDU (1400c),
which is against the blockwise feature's design.

Adjusting those tests, seems to fix the problem.

Bug: https://jira.iotivity.org/browse/IOT-1902
Change-Id: I0aa6d97de0987d13c54b426b537b5d83235517b7
Thanks-to: Pawel Winogrodzki <pawelwi@microsoft.com>
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17827
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Pawel Winogrodzki <pawelwi@microsoft.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>