[IOT-1785] Use a CBOR map for roles in ACL resource
Previous solution used a temporary custom string format to
differentiate UUID-based and role-based subjects. Change this
to using a proper map inside the CBOR for a role, and
differentiate based on whether the subject value is a text string
(in which case it's a UUID type) or a map (in which case a role).

Plug another couple memory leaks found along the way.

Change-Id: Ic592d3538685c9a9ce664619099cfe2ac4889033
Signed-off-by: Kevin Kane <kkane@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/17927
Reviewed-by: Alex Kelley <alexke@microsoft.com>
Reviewed-by: Pawel Winogrodzki <pawelwi@microsoft.com>
Reviewed-by: Greg Zaverucha <gregz@microsoft.com>