build: Download gtest 1.7.0 from upstream source
Fedora mirror was used before, as fallback to googlecode repo.
Then upstream relocated to github, and changed (its tarball form),
so it was changed in iotivity's Sconscripts,
but not in that prep.sh scriptlet.

Now everything is aligned to upstream,
and will not be downloaded twice in different forms
(2d one is scons' puller).

Bug: https://jira.iotivity.org/browse/IOT-1745
Change-Id: I0778499e96121124df0da536623155d9b049bb2f
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/18127
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>