[IOT-1891] [RI] Issue fix OBSERVE Cancellation High QoS
When cancel observe is requested by client, the callback should return
a sequence number greater than MAX_SEQUENCE_NUMBER, and with no observe
header option.
Currently it is returning 0 which is not correct.So stack has been
modified to return a higher sequence number with the callback
for cancel observe confirmed response.

Change-Id: I85ea87454e3ca96f22fe7e414189ec0be7ea8dd8
Signed-off-by: arya.k <arya.kumar@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19651
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>