Normalize scons calls to SConscript()
There have been several different styles used in the project to construct
the path to reading other scons files.  Paths used internally by scons
can use the shorthand of a leading # sign, which can also sort out the
path style, so techniques like fetching SRC_DIR from the environment,
and joining path elements together with os.path.join() are not needed.
For greater readability, the calling style is normalized with this change.

Adapted to comments on build_common/external_builders.scons, and
improved some error messages in that file.

Other paths, such as CPPPATH, can also be modified this way, but
that is left to a different patch series.

Change-Id: Id9413353841a4b19a87981fbdb12a11a931356b2
Signed-off-by: Mats Wichmann <mats@linux.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19673
Reviewed-by: George Nash <george.nash@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>