[IOT-1998] Set remote ID
Since "f7e8c4b PSK check before ciphersuite selection" PSK
ciphersuites will only be negotiated if the local peer has
a credential with the remote peer's UUID. Make sure the UUID
is present in the endpoint information used when calling
OCDoRequest.

In the SSL adapter, when setting up a connection, if no
ciphersuites are usable based on the credentials available, fail.

Change-Id: I64db80379e8055e35e051dd5b90191625f1d5033
Signed-off-by: Greg Zaverucha <gregz@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19123
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>