[IOT-2247] [IOT-2308] Clean up provisioning/sample/*.json files
The goal is to make the .json more readable an remove confusing extra
values.

I tried to leave them functionally identical, with
the exception that I merged the /doxm (Read, Update) and /roles
(Full Control) ACEs into a single ACE with permission 14
(Read, Update, Delete) which is sufficient for both /doxm and /roles.

I verified that at least JustWorks and Random PIN files work,
as well as security unittests.

Change-Id: Id4e7c94a7ce941b5e396b61096b89732aab1be4e
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19923
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Kevin Kane <kkane@microsoft.com>