Modify observer list to be belong to resource
Currently all observer info is managed in a global list in ocobserve.c.
Conceptually, observers should be belong to each resource which the
observer is interested in.
It is more intuitive and easy to manage observers per each resource.
And the function codes in ocobserve could be simple by optimization
after this patch merged.

Change-Id: Ic76816a2729d3fd1715ba5f3f1e5de4253167909
Signed-off-by: Jaewook Jung <jw0213.jung@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20397
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>