Build warning cleanup policyengine.cpp unit test
This cleans up [-Wcomment] warning from gcc.

This warning is given if comments end in the '\',
i.e. the line continuation character, because this
will cause the comment to be wrapped to the next
line of code.

This removes the error without changing the code.

Change-Id: I1323d4236f73985613b4acadc5655863310ab9f2
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20489
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
(cherry picked from commit 1e70a844529a8810975f091c1869cef92946f8d2)
Reviewed-on: https://gerrit.iotivity.org/gerrit/20595
Reviewed-by: Randeep Singh <randeep.s@samsung.com>