Reformat all of the build_common SConscripts to be more Pythonic
There are not intended to be any functional changes.
The help variable definitions are split into multi-line,
in anticipation of later improving the descriptions and using
keyword-argument style to make the purpose of other args more clear.

Also included two files at the top level in this set,
as they seem to have the same intent as build_common.

In build_common/android/SConscript, an identical 4-line stanza, beginning:
  elif cmp(flags[0:9], 'PLATFORM=') == 0:  # get target platform: android-x
appears three times, the second and third were dropped in this change.

Change-Id: I0077e7350754fef25e75621fdb4a41cfcf28ef2a
Signed-off-by: Mats Wichmann <mats@linux.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20545
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: George Nash <george.nash@intel.com>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>