build: Reformat the service SConscripts
[Mats Wichmann]

These reformattings are not supposed to change any behavior,
but they are not all purely whitespace. For example, consecutive
lines of Append(LIBS=  or Append(CPPPATH= were collapsed into
one call with multiple entries in the list.

A few comments were beefed up, as were occasional informative messages.

Import of run_test moved to top (Python style), except the two scripts
where the unittest code is not isolated to a subdirectory - the import
has to be protected so it is only called for targets that support
the tests - the import fails on Tizen for some reasons.  Change this
when separate patches to isolate unit test building/running are merged.

[Philippe Coval]

Ported to 1.3-rel from master

Conflicts:
	service/easy-setup/enrollee/unittests/SConscript
	service/easy-setup/mediator/richsdk/unittests/SConscript
	service/notification/cpp-wrapper/unittest/SConscript
	service/notification/unittest/SConscript
	service/resource-container/unittests/SConscript
	service/resource-encapsulation/src/common/SConscript
	service/resource-encapsulation/src/serverBuilder/SConscript
	service/resource-encapsulation/unittests/SConscript

Change-Id: Ifaf3eb2d837e773a06e59b50132f0d3c8881fdd8
Signed-off-by: Mats Wichmann <mats@linux.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20553
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: George Nash <george.nash@intel.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Bug: https://jira.iotivity.org/browse/IOT-1745
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21001
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>