IOT-1828 Implement ocevent
This change implements ocevent object in c_common, which will be used
by a later change in OCStack to remove the responsibility of calling
OCProcess from applications so that they don't have to call it in an
endless loop.

Note that the patch requires the oc_mutex_assert_owner fix change ID:
I0d60188af9bce866f5ba18800292204c78daba12.

Change-Id: I0d60188af9bce866f5ba18800292204c78daba10
Signed-off-by: Way Vadhanasin <wayvad@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/19311
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>