disable BLE Multicast for Tizen Platform.
once BLE is not supported multicast.
and also current codes related multicast is not working.
(since it has some methods which has wrong parameter)
and code related tizen 3.0 will be updated soon,
it is better to remove the codes first.
otherwise it will causes a big problem related build error

Change-Id: I6dfb29668450b628e98438f3631609c6778d6ffb
Signed-off-by: jihwan.seo <jihwan.seo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20751
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>