Remove 'using namespace' keyword from JNI headers
The use of the 'using namespace OC' was causing an
ambiguous symbol between the C++ OC::OCResource and
the C OCResource when including ocpayload.h.

It is bad practice to use 'using namespace' key words in
header files because any code that includes that header
also imports the namespace, whether they want it or not.

This defeats the purpose of namespaces and can cause name
collisions.

This is part of the work for IOT-2405

Bug: https://jira.iotivity.org/browse/IOT-2405
Change-Id: I8e7e01b1bf645b6333166bd4624a82e9f0345bf8
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20767
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>