[IOT-2405] Enable building generic java in Windows
Changed the name of the iotivity-linux folder to
iotivity-java. Update the build scripts to account
for the name change of the folder. The name of the
output jar file is no longer iotivity-linux.jar it
is now iotivity.jar.

The JNI code links against the static librarys not
the shared librarys for oc, oc_logger, octbstack,
and connectivity_abstraction. Not all libraries are
have a dll built for windows, they must be statically
linked. This insures the Java code is identical for
all platforms. This also greatly reduces the number
of dependencies for the java code on native libraries.
The code now only needs to import ocstack-jni for the
IoTivity stack and ca-interface for security.

Other small updates to address compiler warnings.

Bug: https://jira.iotivity.org/browse/IOT-2405
Change-Id: I98392a6d0c2a5299746fa7dd6540724b0f5a4bc8
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21061
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>