Reformat resource/csdk SConscripts to be more Pythonic 4/4
These are the resource/csdk/security subset

There are not intended to be any functional changes, with a minor exception:
in resource/csdk/security/provisioning/sample/SConscript, the samples are
now assigned to an appropriately constructed list, which is extended
in the case of MULTIPLE_OWNER, this list is then assigned an Alias.
This is done to avoid having two parallel lists and needing to keep
in sync.

Header lines with doubled comment styles were changed to just Python,
that is, change this style (quoted so git doesn't strip the lines):

"# // Copyright 2015 Samsung Electronics All Rights Reserved."

to this:

"# Copyright 2015 Samsung Electronics All Rights Reserved."

There is no consistently "pretty" way to fold long lines,
the choices were made by a tool (yapf) but occasionally overridden
to be a little more consistent

The test harness is now consistently imported as:
  from tools.scons.RunTest import run_test
rather than import * - should normally avoid importing whole
namespace unqualified unless it is really needed

Change-Id: If1876589b2cad83c1c4795dae5e3ddc59c78fca1
Signed-off-by: Mats Wichmann <mats@linux.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/20805
Reviewed-by: George Nash <george.nash@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>