IOT-2490 Fix JNI build using WITH_CLOUD=1 option
This build failure was a result of a recent refactor to
remove the `using namespace OC` from the header files.
Importing the name was casing name space collisions
between C and C++ code used in the JNI files when built
in Windows OS.

The WITH_CLOUD build option was missed in the refactor.

Bug: https://jira.iotivity.org/browse/IOT-2490

Change-Id: I89b4c35a5a312d3e09a69016e43762fb01359abe
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21347
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Jay  Sharma <jay.sharma@samsung.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>