build: prevent unused-parameter warning
the connectivity bt_le_adapter has auto generates files
  - bluez-glue.c
  - object_manager-glue.c

The files are included when building the connectivity_abstraction
library. This error is only disabled when bt_le_adapter is being
built. We don't want to hide the unused-parameter warning. This
is the simplest way to hide the warning without large changes
in the build.

Change-Id: Ieca1412ab744b7fccf408a80f66ef0add1a91fa6
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21429
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Reviewed-by: Mats Wichmann <mats@linux.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>