Make OCServerRequest rb-tree items linked lists
If a notification is issued in response to an incoming observe request
then the CAToken_t of the OCServerRequest generated to perform the
notification has the same value as the CAToken_t of the incoming
observe request. Since requests are stored in a red-black tree keyed on
the CAToken_t there can be no duplicates. Thus, the OCServerRequest
corresponding to the notificatin is never added to the OCServerRequest
tree, and when the notification is completed the OCServerRequest
representing the original incoming observe request is removed. When the
application then proceeds to send a response to the original incoming
observe request via OCDoResponse(), the request is not found.

The solution is to allow the OCServerRequest red-black tree nodes to be
linked lists. Thus, for each CAToken_t there may be multiple entries at
that particular node. Only when the last entry at a node is removed is
the node itself removed from the tree.

Making this change also obviates the need to retrieve an
OCServerRequest by handle, since the handle is just the OCServerRequest
itself cast as an OCRequestHandle.

Bug: https://jira.iotivity.org/browse/IOT-2417
Change-Id: I9d39b9483d4dc3f6035ff03ad1962f4a586a6331
Signed-off-by: Gabriel Schulhof <gabriel.schulhof@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21239
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>