Catch up 1.3-rel on plugins reformats
Some sconscripts got missed pulling reformats over from master.
This changeset is the scripts in plugins - these six scripts will
now be identical to their counterpart in master.

Note this does pick up the change not to locally set BUILD_DIR
into LIBPATH and to not set TB_LOG into CPPDEFINES as those are
done on a global level already.

Bug: https://jira.iotivity.org/browse/IOT-1745
Change-Id: Ib793b7eeaeda1d380b805153e82cffade7e65c80
Signed-off-by: Mats Wichmann <mats@linux.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21463
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>