notification: Sync JSON and CBOR files
To avoid local changes on build,
let's commit the updated .dat files (using json2cbor),

Reminding, It's developer dutty to commit those files on each changes
on json source or any change that produce a different result.
This way we're not leaving inconsistant binary files in tree.

JSON file was not modified because it is already in UNIX format

Note: those files may need to be updated to support current ACL
(and commited by security team, check status of json and dat files)

Bug: https://jira.iotivity.org/browse/IOT-2310
Change-Id: Ie452375af9418e0cbf3513e5334ce15cb4994a86
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21847
Reviewed-by: George Nash <george.nash@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>