[IOT-2542] Wrap bt_le_files with gcc pragma
Wrap the files that are generated using the gdbus-codegen
tool with the GCC diagnostic ignore -Wunused-parameter
pragma to prevent this code producing the unused-parameter
warning.

If transports was set to ALL the build script was calling
build for all of the connectivity transports twice. This
caused SCons to print a warning that two builds were
specifying the same build target. To fix the extra
unneeded call to the build script was removed.

Bug: https://jira.iotivity.org/browse/IOT-2542
Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I997506b43099425718a7f04f9c66be49a49dc717
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21641
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>