build: removed the -Wno-comment build flag
This build flag was added because of the way the
copyright was writen mixed single line and multi-line
comments in many of the files. Resulting in multiple
-Wcomment warnings.  Since the extra forward slashes
`//` in the copyright were not needed. The copyright
was updated so the warning could be re-enabled.

The major reason for the comment warning is to make
sure users are not inadvertantly breaking the
compilers comment parsing. Its possible to
inadvertantly comment out more code than is intended.

With modern editors this is normally not an issue.
It was still easy to fix, so the warning was removed
from the list of ignored warnings.

Bug: https://jira.iotivity.org/browse/IOT-2539
Change-Id: I605775b498726e7fb3202473a9b8169b7c8c945e
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21583
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>