Convert href, if, rt on links payload
Previously OCPayloadConvert logic in sever side only convert uri,
types and interfaces for the top level element.
Even if a RepPayload exists in the RepPayload, uri and so on are not
converted.

This patch move the uri, type and interface convert logic into 
OCConvertSingleRepPayload which is recrsively called function.

Previous code skipped to convert href in case array payload size is 1
on the top level payload. This logic should be removed at least
considering collection resource with linked-list interface request.
Furthermore, this skip logic is not mandatory code for certification
as of now.

Change-Id: I17b1cc02aed37e9af55126d3af2d28310a3ae733
Signed-off-by: uzchoi <uzchoi@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/21629
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Todd Malsbary <todd.malsbary@intel.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>